ng-demo [ref] renamed /lu-popover to /popover
