+++ packages/ng/demo/app/app.module.ts
@@ -14,7 +14,7 @@ import { appRoutes } from './app.router';
 import { DemoDateRangePickerModule } from './date-range-picker/date-range-picker.module';
 import {DemoFormlyModule} from './formly/formly.module';
 import {DemoEmptyModule} from './empty/empty.module';
-import { DemoPopoverModule } from './lu-popover/lu-popover.module';
+import { DemoPopoverModule } from './popover/popover.module';
 import { DemoAnimationsModule } from './animations/animations.module';
 import { DemoApiModule } from './api/api.module';
