+++ packages/ng/demo/app/navigation/navigation.router.ts
@@ -2,7 +2,7 @@ import { Routes } from '@angular/router';
 import { dateRangePickerRoutes } from '../date-range-picker/date-range-picker.router';
 import { formlyRoutes } from '../formly/formly.router';
 import { emptyRoutes } from '../empty/empty.router';
-import { luPopoverRoutes } from '../lu-popover/lu-popover.router';
+import { popoverRoutes } from '../popover/popover.router';
 import { animationsRoutes } from '../animations/animations.router';
 import { apiRoutes } from '../api/api.router';
 import { userRoutes } from '../user/user.router';
@@ -10,7 +10,7 @@ export const featuresRoutes = [
 	...dateRangePickerRoutes,
 	...formlyRoutes,
 	...emptyRoutes,
-	...luPopoverRoutes,
+	...popoverRoutes,
 	...animationsRoutes,
 	...apiRoutes,
 	...userRoutes,
