+++ packages/ng/demo/app/popover/basic/basic.html
@@ -1,0 +1,11 @@
+<div class="button" [LuPopoverTriggerFor]="popover">Popover on click</div>
+<div class="button" [LuPopoverTriggerFor]="userDetails">Popover on hover, overlapping</div>
+
+<div class="spacer" style="height: 300px"></div>
+<lu-popover #popover trigger-on="click" [overlap-trigger]="false" [close-on-click]="false" [position]="'after'" [alignment]="'top'">
+	<lu-user-tile class="user-tile" [user]="user"></lu-user-tile>
+</lu-popover>
+
+<lu-popover #userDetails trigger-on="hover" [overlap-trigger]="true">
+	<lu-user-tile class="user-tile" [user]="user"></lu-user-tile>
+</lu-popover>
