+++ packages/ng/demo/app/popover/popover.component.html
@@ -1,5 +1,5 @@
 <demo-api-docs directive="LuPopoverComponent"></demo-api-docs>
 <h3>Examples</h3>
 <demo-example-box [snippets]="snippets" demo="basic">
-	<demo-lu-popover-basic></demo-lu-popover-basic>
+	<demo-popover-basic></demo-popover-basic>
 </demo-example-box>
